42-FE: Move "forgot password" link into login form
