+++ Okane.Client/src/shared/pages/LoginPage.spec.ts
@@ -2,9 +2,6 @@
 import LoginPage from '@shared/pages/LoginPage.vue'
 import LoginForm from '@features/auth/components/LoginForm.vue'
 
-import { appRoutes } from '@shared/services/router/router'
-import { AUTH_COPY } from '@features/auth/constants/copy'
-
 const mountComponent = getMountComponent(LoginPage, {
   withQueryClient: true,
   withRouter: true,
@@ -13,10 +10,4 @@ const mountComponent = getMountComponent(LoginPage, {
 test('renders a LoginForm', () => {
   const wrapper = mountComponent()
   expect(wrapper.findComponent(LoginForm).exists()).toBe(true)
-})
-
-test('renders a "forgot password" link', () => {
-  const wrapper = mountComponent()
-  const link = wrapper.findByText('a', AUTH_COPY.FORGOT_PASSWORD)
-  expect(link.attributes('href')).toBe(appRoutes.sendResetPasswordEmail.buildPath())
 })
